fix: strict warnings, OpenSSL const, CI test ordering
- addrbook.c: cast int64 to size_t before division to match %zu format
- key.c: cast string literal to (char *) for OpenSSL 3 OSSL_PARAM API
- ci.yml: run make check after debug build to avoid sanitizer link mismatch

+++ core/addrbook.c
@@ -240,7 +240,7 @@ addrbook_open(struct config *config,
       char *name = file_getname(book->filename);
 
       log_warn(LGPFX" reading file %s -- %s -- %zu addrs.\n",
-              name, s, size / sizeof(btc_msg_address));
+              name, s, (size_t)size / sizeof(btc_msg_address));
       free(name);
       free(s);
    }
